Structuration of calendar view
