+++ composeApp/src/commonMain/kotlin/com/project/lifeos/ui/view/calendar/CalendarDataSource.kt
@@ -1,7 +1,8 @@
-package com.project.lifeos.ui.screen
+package com.project.lifeos.ui.view.calendar
 
 import java.time.DayOfWeek
 import java.time.LocalDate
+import java.time.format.DateTimeFormatter
 import java.time.temporal.ChronoUnit
 import java.util.stream.Collectors
 import java.util.stream.Stream
@@ -46,4 +47,21 @@ class CalendarDataSource {
         isToday = date.isEqual(today),
         date = date,
     )
+}
+
+data class CalendarUiModel(
+    val selectedDate: Date, // the date selected by the User. by default is Today.
+    val visibleDates: List<Date> // the dates shown on the screen
+) {
+
+    val startDate: Date = visibleDates.first() // the first of the visible dates
+    val endDate: Date = visibleDates.last() // the last of the visible dates
+
+    data class Date(
+        val date: LocalDate,
+        val isSelected: Boolean,
+        val isToday: Boolean
+    ) {
+        val day: String = date.format(DateTimeFormatter.ofPattern("E")) // get the day by formatting the date
+    }
 }
